--- a/parlai/agents/bart/__init__.py
+++ b/parlai/agents/bart/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.