--- a/parlai/agents/bart/bart.py
+++ b/parlai/agents/bart/bart.py
@@ -8,6 +8,10 @@ BART: Denoising Sequence-to-Sequence Pre-training for
 Natural Language Generation, Translation, and Comprehension
 
 See https://arxiv.org/abs/1910.13461.
+
+The BART agent can be instantiated as simply `-m bart`,
+however it is recommended to specify `--init-model zoo:bart/bart_large/model`
+or `-mf zoo:bart/bart_large/model` to ensure correct dictionaries are saved.
 """
 import os
 import torch
@@ -18,6 +22,7 @@ from parlai.agents.transformer.transformer import TransformerGeneratorAgent
 from parlai.core.agents import compare_init_model_opts
 from parlai.core.message import Message
 from parlai.core.opt import Opt
+from parlai.core.params import ParlaiParser
 from parlai.core.torch_agent import Batch, History
 from parlai.scripts.convert_fairseq_to_parlai import ConversionScript
 from parlai.utils.typing import TShared
@@ -28,65 +33,54 @@ class BartAgent(TransformerGeneratorAgent):
     """
     BART Agent.
 
-    Relies on BART model implemented in fairseq.
+    Relies on the BART model implemented in fairseq.
 
-    Can initialize with BART, or can handle
+    If you have a fine-tuned BART model from fairseq, you can specify the
+    `--init-fairseq-model` arg, which will convert your fine-tuned model
+    to a ParlAI model.
     """
+
+    @classmethod
+    def add_cmdline_args(cls: "BartAgent", argparser: ParlaiParser):
+        """
+        Override to add init-fairseq-model arg.
+        """
+        TransformerGeneratorAgent.add_cmdline_args(argparser)
+        group = argparser.add_argument_group('Bart Args')
+        group.add_argument(
+            '--init-fairseq-model',
+            type=str,
+            default=None,
+            help='fairseq checkpoint for bart',
+        )
 
     def __init__(self, opt: Opt, shared: TShared = None):
         if not shared:
-            if not opt.get('converting'):
-                download(opt['datapath'])
-                opt['init_model'] = os.path.join(opt['datapath'], 'models/bart_models/bart.large/model')
-            if opt.get('init_fairseq_model'):
-                opt = self._convert_model(opt)
-            opt.update(BART_ARGS)
-            compare_init_model_opts(opt, opt)
+            opt = self._initialize_bart(opt)
         super().__init__(opt, shared)
 
-    def build_model(self, states=None):
+    def _initialize_bart(self, opt: Opt) -> Opt:
         """
-        Build and return model.
+        Download and convert BART pre-trained models.
+
+        Additionally, convert `init-fairseq-model` if necessary.
+
+        :param opt:
+            ParlAI-parsed options
+
+        :return opt:
+            return opt with BART-specific args.
         """
-        model = BartModel(self.opt, self.dict)
-        if self.opt['embedding_type'] != 'random':
-            self._copy_embeddings(
-                model.encoder.embeddings.weight, self.opt['embedding_type']
+        if not opt.get('converting'):
+            download(opt['datapath'])
+            opt['init_model'] = os.path.join(
+                opt['datapath'], 'models/bart/bart_large/model'
             )
-        return model
-
-    def _get_initial_decoder_input(self, bsz, beam_size, dev):
-        """
-        Override to seed decoder with EOS token.
-        """
-        return (
-            torch.LongTensor([self.END_IDX]).expand(bsz * beam_size, 1).to(dev)
-        )
-
-    def _generate(
-        self,
-        batch: Batch,
-        beam_size: int,
-        max_ts: int,
-        prefix_tokens: Optional[torch.LongTensor] = None
-    ):
-        """
-        Override to set prefix_tokens.
-
-        For bart pretraining, a bos token was added to the input.
-        so basically the input to encoder is:
-
-        <bos> seq <eos>
-
-        input to decoder:
-        <eos> <bos> seq
-
-        target is:
-        <bos> seq <eos>
-        """
-        if batch.text_vec is not None:
-            prefix_tokens = batch.text_vec.new_zeros((len(batch.text_vec), 1)).fill_(self.START_IDX)
-        return super()._generate(batch, beam_size, max_ts, prefix_tokens)
+        if opt.get('init_fairseq_model'):
+            opt = self._convert_model(opt)
+        opt.update(BART_ARGS)
+        compare_init_model_opts(opt, opt)
+        return opt
 
     def _convert_model(self, opt: Opt) -> Opt:
         """
@@ -100,34 +94,32 @@ class BartAgent(TransformerGeneratorAgent):
         """
         model_name = os.path.split(opt['init_fairseq_model'])[-1]
         args = CONVERSION_ARGS.copy()
+
         args['input'] = [opt['init_fairseq_model']]
         if opt.get('model_file') and not os.path.exists(opt['model_file']):
             args['output'] = opt['model_file']
         else:
-            args['output'] = os.path.join(opt['datapath'], 'models/converted_fairseq_models/', model_name)
+            args['output'] = os.path.join(
+                opt['datapath'], 'models/converted_fairseq_models/', model_name
+            )
+
         ConversionScript.main(**args)
         opt['init_model'] = args['output']
         return opt
 
-    @classmethod
-    def add_cmdline_args(cls, argparser):
+    def build_model(self) -> BartModel:
         """
-        Override to add one arg.
+        Build and return model.
         """
-        TransformerGeneratorAgent.add_cmdline_args(argparser)
-        group = argparser.add_argument_group('Bart Args')
-        group.add_argument(
-            '--init-fairseq-model',
-            type=str,
-            default=None,
-            help='fairseq checkpoint for bart',
-        )
+        model = BartModel(self.opt, self.dict)
+        if self.opt['embedding_type'] != 'random':
+            self._copy_embeddings(
+                model.encoder.embeddings.weight, self.opt['embedding_type']
+            )
+        return model
 
     def _set_text_vec(
-        self,
-        obs: Message,
-        history: History,
-        truncate: Optional[int],
+        self, obs: Message, history: History, truncate: Optional[int]
     ) -> Message:
         """
         Override to prepend start token and append end token.
@@ -146,8 +138,44 @@ class BartAgent(TransformerGeneratorAgent):
                 [
                     vec.new_tensor([self.dict[self.dict.start_token]]),
                     vec,
-                    vec.new_tensor([self.dict[self.dict.end_token]])],
-                0
-            )
+                    vec.new_tensor([self.dict[self.dict.end_token]]),
+                ],
+                0,
+            ),
         )
         return obs
+
+    def _get_initial_decoder_input(self, bsz: int, beam_size: int, dev: str):
+        """
+        Override to seed decoder with EOS token.
+
+        See docstring for `BartAgent._generate` for more details.
+        """
+        return torch.LongTensor([self.END_IDX]).expand(bsz * beam_size, 1).to(dev)
+
+    def _generate(
+        self,
+        batch: Batch,
+        beam_size: int,
+        max_ts: int,
+        prefix_tokens: Optional[torch.LongTensor] = None,
+    ):
+        """
+        Override to set prefix_tokens.
+
+        For bart pretraining, a bos token was added to the input.
+
+        input to encoder:
+        <bos> seq <eos>
+
+        input to decoder:
+        <eos> <bos> seq
+
+        target is:
+        <bos> seq <eos>
+        """
+        if batch.text_vec is not None:
+            prefix_tokens = batch.text_vec.new_zeros((len(batch.text_vec), 1)).fill_(
+                self.START_IDX
+            )
+        return super()._generate(batch, beam_size, max_ts, prefix_tokens)