--- a/parlai/agents/bart/bart.py
+++ b/parlai/agents/bart/bart.py
@@ -17,6 +17,7 @@ import os
 import torch
 from typing import Optional, Any
 
+from parlai.agents.bart.convert_fairseq_to_parlai import ConversionScript
 from parlai.agents.bart.modules import BartModel
 from parlai.agents.transformer.transformer import TransformerGeneratorAgent
 from parlai.core.agents import compare_init_model_opts
@@ -24,7 +25,6 @@ from parlai.core.message import Message
 from parlai.core.opt import Opt
 from parlai.core.params import ParlaiParser
 from parlai.core.torch_agent import Batch, History
-from parlai.scripts.convert_fairseq_to_parlai import ConversionScript
 from parlai.utils.typing import TShared
 from parlai.zoo.bart.build import download, CONVERSION_ARGS, BART_ARGS
 
@@ -53,7 +53,7 @@ class BartAgent(TransformerGeneratorAgent):
             default=None,
             help='fairseq checkpoint for bart',
         )
-        argparser.set_defaults(dict_maxexs=0)  # skip building dictionary
+        argparser.set_defaults(dict_tokenizer='gpt2')
 
     def __init__(self, opt: Opt, shared: TShared = None):
         if not shared: