@emilydinan 2020-06-24
Re: not building dictionary. You can do a hack like this: https://github.com/facebookresearch/ParlAI/blob/ce0caf63ae951cca9d4c52b4b4b4d1632bcdf980/parlai/agents/hugging_face/gpt2.py#L229

(Just set max_exs to 0 so it's not built)