@stephenroller 2020-06-24
`tests/nightly/gpu/test_bart.py::TestBartModel::test_bart SKIPPED         [ 33%]`
You need to add fairseq into the nightly installs. Same place we install nltk i think.