@klshuster 2020-06-25
i moved the script to `parlai/agents/bart/` for now